requirements.txt
Django==3.1.2
djangorestframework>=3.11.1,<3.11.2
psycopg2-binary>=2.8.0,<2.8.7
drf-yasg>=1.17.1,<1.17.2